+++ src/main/java/com/myfactory/SBootWebProject/dto/FacturaLineaDTO.java
@@ -6,17 +6,16 @@ public class FacturaLineaDTO implements Serializable {
 	
 	private static final long serialVersionUID = 1L;
 
-	private Long CANTIDAD;
+	private Float CANTIDAD;
 	private String CONCEPTO;
-	private Long POR_IVA;
+	private Float POR_IVA;
 	private Float IMP_LIN_FACTURA;
-	private Long POR_DESCUENTO;
+	private Float POR_DESCUENTO;
 	
-	
-	public Long getCANTIDAD() {
+	public Float getCANTIDAD() {
 		return CANTIDAD;
 	}
-	public void setCANTIDAD(Long cANTIDAD) {
+	public void setCANTIDAD(Float cANTIDAD) {
 		CANTIDAD = cANTIDAD;
 	}
 	public String getCONCEPTO() {
@@ -25,10 +24,10 @@ public class FacturaLineaDTO implements Serializable {
 	public void setCONCEPTO(String cONCEPTO) {
 		CONCEPTO = cONCEPTO;
 	}
-	public Long getPOR_IVA() {
+	public Float getPOR_IVA() {
 		return POR_IVA;
 	}
-	public void setPOR_IVA(Long pOR_IVA) {
+	public void setPOR_IVA(Float pOR_IVA) {
 		POR_IVA = pOR_IVA;
 	}
 	public Float getIMP_LIN_FACTURA() {
@@ -37,10 +36,10 @@ public class FacturaLineaDTO implements Serializable {
 	public void setIMP_LIN_FACTURA(Float iMP_LIN_FACTURA) {
 		IMP_LIN_FACTURA = iMP_LIN_FACTURA;
 	}
-	public Long getPOR_DESCUENTO() {
+	public Float getPOR_DESCUENTO() {
 		return POR_DESCUENTO;
 	}
-	public void setPOR_DESCUENTO(Long pOR_DESCUENTO) {
+	public void setPOR_DESCUENTO(Float pOR_DESCUENTO) {
 		POR_DESCUENTO = pOR_DESCUENTO;
 	}
